Refactor 3D Mask Editor to decouple UI and state management
